* g++.dg/4425360.C: Tweak for -m64.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/apple-local-200502-branch@112189 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/g++.dg/4425360.C
@@ -27,7 +27,7 @@ int main (int argc, char * const argv[])
         }
         packet->next = 0;
         /* printf("in loop %d         head = 0x%X    next = 0x%X\n", count, (unsigned int)head, (unsigned int)head->next); */
-        printf("", count, (unsigned int)head, (unsigned int)head->next);
+        printf("", count, head, head->next);
     } while (++count < num);
     if (!head->next)
       abort ();
